(TODO box test) Use Intrinsics.checkNotNull and throw NPE in !! operator generation
This method was introduced in c204e8fc67 "just in case" and was never
used. Therefore we're free to change its semantics and use it in all new
generated code (with API version >= 1.4), without even the
isApiVersionAtLeast check in inline functions. Since we agreed to change
the type of thrown exceptions to java.lang.NPE in KT-22275, call a new
method throwJavaNpe now which throws that exception instead of KNPE.

Note that the additional method that takes an exception message is still
unused and exists just in case we need to use it in the future.

 #KT-22275 In Progress